Support for decryption using Shaka packager and for live streams.
The --decryption-application commandline option makes it possible to select Shaka packager instead
of mp4decrypt for decryption of media streams with ContentProtection.

The --enable-live-streams and --force-duration options enable preliminary support for dynamic
manifests / live media streams. This will only work with a small subset of "pseudo-live" streams, as
explained in the README.