CI: change release action to avoid use of cross-rs.
The GitHub action taiki-e/upload-rust-binary-action depends on cross-rs, which
in its default configuration is using podman images that don't include a C
compiler, and therefore fail to build the pssh-box crate (which builds the
protoc compiler from source at build time).

This reduces the number of platforms for which we are able to provide release
binaries (currently no aarch64-unknown-linux-gnu nor aarch64-apple-darwin).